# docs/api_doc/swagger.json
{"openapi":"3.0.1","info":{"title":"'@project.name@'","description":"'@project.description@'","contact":{"name":"Joe Doe","email":"dumy@dumy"},"version":"'@project.version@'"},"servers":[{"url":"http://localhost:8080","description":"'@project.description@'"}],"paths":{"/reservations":{"put":{"tags":["reservation-controller"],"operationId":"updateReservation","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReservationDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ReservationDto"}}}}}},"post":{"tags":["reservation-controller"],"operationId":"createReservation","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReservationDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/flights":{"post":{"tags":["flights-controller"],"operationId":"postFlight","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/FlightDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/FlightDto"}}}}}}},"/reservations/{id}":{"get":{"tags":["reservation-controller"],"operationId":"getReservationById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ReservationDto"}}}}}}},"/ping":{"get":{"tags":["ping-controller"],"summary":"Check healthy of the app","description":"If the app it's alive response pong","operationId":"pong","responses":{"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorApi"}}}},"200":{"description":"Successful operation","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/flights/{id}":{"get":{"tags":["flights-controller"],"operationId":"getFlightById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/FlightDto"}}}}}}}},"components":{"schemas":{"PassengersDto":{"type":"object","properties":{"name":{"type":"string"},"seat":{"type":"string"}}},"ReservationDto":{"type":"object","properties":{"id":{"type":"string"},"status":{"type":"string"},"flight":{"type":"string"},"passengers":{"type":"array","items":{"$ref":"#/components/schemas/PassengersDto"}}}},"AirportDto":{"type":"object","properties":{"name":{"type":"string"},"code":{"type":"string"},"location":{"type":"string"}}},"FlightDto":{"type":"object","properties":{"id":{"type":"string"},"aircraft":{"type":"string"},"departure":{"type":"string","format":"date-time"},"airport":{"$ref":"#/components/schemas/AirportDto"},"seat_map":{"type":"array","items":{"$ref":"#/components/schemas/SeatDto"}}}},"SeatDto":{"type":"object","properties":{"seat":{"type":"string"},"status":{"type":"string"}}},"ErrorApi":{"type":"object","properties":{"timestamp":{"type":"string"},"status":{"type":"integer","format":"int32"},"error":{"type":"string"},"message":{"type":"string"}}}}}}