# src/main/java/ar/edu/utn/frc/tup/lc/iv/dtos/AirportDto.java
package ar.edu.utn.frc.tup.lc.iv.dtos;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@AllArgsConstructor
@NoArgsConstructor
@Builder
public class AirportDto {
    private String name;
    private String code;
    private String location;
}